+++ src/main/kotlin/com/ajaxproject/telegrambot/bot/configuration/TelegramBotsConfiguration.kt
@@ -2,6 +2,7 @@ package com.ajaxproject.telegrambot.bot.configuration
 
 import com.ajaxproject.telegrambot.bot.BankInfoBot
 import com.ajaxproject.telegrambot.bot.properties.BotProperties
+import com.ajaxproject.telegrambot.bot.properties.MonobankProperties
 import com.ajaxproject.telegrambot.bot.properties.TextProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
@@ -10,7 +11,7 @@ import org.telegram.telegrambots.meta.TelegramBotsApi
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession
 
 @Configuration
-@EnableConfigurationProperties(BotProperties::class, TextProperties::class)
+@EnableConfigurationProperties(BotProperties::class, TextProperties::class, MonobankProperties::class)
 class TelegramBotsConfiguration {
     @Bean
     fun telegramBotsApi(property: BankInfoBot): TelegramBotsApi =
